Dedupe withdraw.js and mint.js code with util.js
Also, fix withdraw.js stateRoot input. It had the second state root
rather than the initial one. This still passed the tests because
withdraw optionally doesn't update the root, since forcing is possible.
