+++ circuits/inputs/mint.json
@@ -5,10 +5,6 @@
     ],
     "private_inputs": [
         {
-            "step_in": [
-                "28923420649458651016230972732093672980259561931925916453795346630536953518599",
-                "8124834938220291251923929260846596950253538988929140787772786994413802905081"
-            ],
             "sender": "0",
             "recipient": "802933809494131860455082925493303288586736066170",
             "leaf_coins": [
@@ -37,10 +33,6 @@
             ]
         },
         {
-            "step_in": [
-                "27899618321255540227580586409989954934950349711406341161637950923207881191857",
-                "5565782154730803119410032728411479820942542911541871014149524394701770769722"
-            ],
             "sender": "0",
             "recipient": "253486562210967126009990789802080859110172940592",
             "leaf_coins": [
