+++ circuits/inputs/withdraw.json
@@ -1,6 +1,6 @@
 {
     "step_in": [
-        "27336385660251599878695835306387567517249988783237154757173305497400454872157"
+        "5949490667649339501861400209195061774766595748318086765647255840405588561701"
     ],
     "private_inputs": [
         {
